refactor: moved the loading of floors onto the map into it's  own file for the viewer page
